ai-service/requirements.txt
Flask==3.0.0
fastapi==0.108.0
uvicorn==0.25.0
python-dotenv==1.0.0
transformers==4.36.0
torch==2.1.0
openai==1.6.0
pydantic==2.5.0
numpy==1.26.0
scikit-learn==1.3.0
